enemyFin, camera
- enemy Finish
- enemy의 bullst fire제어는 전부 enemyManager에서.
- camera 이동

+++ Project_MiniGame_181227/scripts/ImageLoader.cpp
@@ -30,7 +30,7 @@ vector<string> ImageLoader::ImageLoad( const char * loadFileName )
 
 	imgInfo = TXTDATA->textLoadPerLine(fileName.c_str(), "//");
 
-	int size = imgInfo.size();
+	int size = (int)imgInfo.size();
 	for ( int ii = 0; ii < size; ++ii )
 	{
 		char str[128] = {};
